+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # EasyMDE Changelog
 All notable changes to EasyMDE will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### Added
+- Check-list toolbar button and the ability to switch between unordered, ordered, and check-list types (Thanks to [@steve3ussr], [#631]).
+
+### Fixed
+- Extra space added before list content when switching between unordered and ordered lists (Thanks to [@steve3ussr], [#631]).
+
 ## [2.20.0] - 2025-03-04
 ### Added
 - Support for `marked` extensions (Thanks to [@codingjoe], [#611], [#514]).
@@ -295,6 +301,7 @@ Project forked from [SimpleMDE](https://github.com/sparksuite/simplemde-markdown
 [#9]: https://github.com/Ionaru/easy-markdown-editor/issues/9
 
 <!-- Linked PRs -->
+[#631]: https://github.com/Ionaru/easy-markdown-editor/pull/631
 [#608]: https://github.com/Ionaru/easy-markdown-editor/pull/608
 [#592]: https://github.com/Ionaru/easy-markdown-editor/pull/592
 [#591]: https://github.com/Ionaru/easy-markdown-editor/pull/591
@@ -445,6 +452,7 @@ Project forked from [SimpleMDE](https://github.com/sparksuite/simplemde-markdown
 [@p1gp1g]: https://github.com/p1gp1g
 [@mayraamaral]: https://github.com/mayraamaral
 [@codingjoe]: https://github.com/codingjoe
+[@steve3ussr]: https://github.com/steve3ussr
 
 <!-- Linked versions -->
 [Unreleased]: https://github.com/Ionaru/easy-markdown-editor/compare/2.20.0...HEAD